% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osim_create_example.R
\name{osim.create.example}
\alias{osim.create.example}
\title{Creates an offsetsim usage example file 'offsetsim-example.R' in the given directory}
\usage{
osim.create.example(outdir = ".", loglevel = INFO)
}
\arguments{
\item{outdir}{the output directory to create the user example file in (default is '.')}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}
}
\description{
Creates an offsetsim usage example file 'offsetsim-example.R' in the given directory
}
