% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_routines.R
\name{osim.output}
\alias{osim.output}
\title{Plots the results of the Offset Simulator run}
\usage{
osim.output(
  user_output_params = NULL,
  simulation_folder = NULL,
  output_type = NULL,
  loglevel = INFO
)
}
\arguments{
\item{user_output_params}{user configured plotting parameters to use}

\item{simulation_folder}{user configured simulation folder to use}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}

\item{run_number}{user defined folder}
}
\description{
Plots the results of the Offset Simulator run
}
