% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collated_realisations.R
\name{plot}
\alias{plot}
\title{Plots the results of the Offset Simulator run}
\usage{
plot(config, loglevel = INFO)
}
\arguments{
\item{config}{user configured plotting parameters to use}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}
}
\description{
Plots the results of the Offset Simulator run
}
