% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_offsets_simulation.R
\name{run}
\alias{run}
\title{Runs the Offset Simulator}
\usage{
run(config = NULL, loglevel = INFO)
}
\arguments{
\item{config}{user configured parameters to use}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}
}
\description{
Runs the Offset Simulator
}
