% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collated_realisations.R
\name{osim.plot}
\alias{osim.plot}
\title{Plots the results of the Offset Simulator run}
\usage{
osim.plot(user_plot_params = NULL, simulation_folder = NULL,
  run_number = NULL, loglevel = INFO)
}
\arguments{
\item{user_plot_params}{user configured plotting parameters to use}

\item{simulation_folder}{user configured simulation folder to use}

\item{run_number}{user defined folder}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}
}
\description{
Plots the results of the Offset Simulator run
}
