% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collated_realisations.R
\name{osim.plot}
\alias{osim.plot}
\title{Plots the results of the Offset Simulator run}
\usage{
osim.plot(plot_params, loglevel = INFO)
}
\arguments{
\item{plot_params}{user configured plotting parameters to use}

\item{loglevel}{logging level to use, for instance futile.logger::INFO}
}
\description{
Plots the results of the Offset Simulator run
}
