% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_offsets_simulation.R
\name{osim.run}
\alias{osim.run}
\title{Runs the Offset Simulator}
\usage{
osim.run(user_global_params = NULL, user_simulation_params = NULL,
  user_simulated_ecology_params = NULL, user_simulation_dynamics = NULL,
  loglevel = WARN)
}
\arguments{
\item{user_global_params}{user configured parameters to use}

\item{user_simulation_params}{user configured parameters to use}

\item{user_simulated_ecology_params}{user configured parameters to use}

\item{loglevel}{logging level to use, for instance futile.logger::WARN}
}
\description{
Runs the Offset Simulator
}
