% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_offsets_simulation.R
\name{osim.run}
\alias{osim.run}
\title{Runs the Offset Simulator}
\usage{
osim.run(config = NULL, loglevel = WARN)
}
\arguments{
\item{config}{user configured parameters to use}

\item{loglevel}{logging level to use, for instance futile.logger::WARN}
}
\description{
Runs the Offset Simulator
}
