% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_offsets_simulation.R
\name{run}
\alias{run}
\title{Runs the Offset Simulator}
\usage{
run(config = "R/initialise_params_defaults.R", loglevel = INFO)
}
\arguments{
\item{config}{user configured parameters to use}
}
\description{
Runs the Offset Simulator
}
